% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{aggregate}
\usage{
\S4method{aggregate}{CoverageExperiment}(x, bin = 1, ...)
}
\arguments{
\item{x}{a \code{CoverageExperiment} or CoverageExperiment` object}

\item{bin}{an integer to bin each assay by. The \code{width} of the
\code{CoverageExperiment} object should be a multiple of \code{bin}.}

\item{...}{ignored}
}
\value{
an \code{CoverageExperiment} object
}
\description{
Bin assays of an \code{CoverageExperiment} to a coarser resolution
}
\examples{
data(ac)
aggregate(ac, bin = 10)
}
