% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{AggregatedCoverage}
\alias{AggregatedCoverage}
\alias{aggregate,CoverageExperiment}
\title{aggregate}
\usage{
\S4method{aggregate}{CoverageExperiment}(x, bin = 1, ...)
}
\arguments{
\item{x}{a \code{CoverageExperiment} object}

\item{bin}{an integer to bin each assay by. The \code{width} of the
\code{AggregatedCoverage} object should be a multiple of \code{bin}.}

\item{...}{ignored}
}
\value{
an \code{AggregatedCoverage} object
}
\description{
Bin coverage contained in a \code{CoverageExperiment} into an
\code{AggregatedCoverage} object.
}
\examples{
data(ce)
aggregate(ce, bin = 10)
}
