% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble-methods}
\alias{as_tibble-methods}
\alias{as_tibble.CoverageExperiment}
\alias{as_tibble.AggregatedCoverage}
\title{as_tibble}
\usage{
\method{as_tibble}{CoverageExperiment}(x, ...)

\method{as_tibble}{AggregatedCoverage}(x, ...)
}
\arguments{
\item{x}{an \code{CoverageExperiment} object}

\item{...}{ignored}
}
\value{
\code{tibble}
}
\description{
Coerce an \code{CoverageExperiment} object into a \code{tibble}
}
\examples{
data(ac)
as_tibble(ac)
}
