% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\alias{expand,CoverageExperiment-method}
\title{expand}
\usage{
\S4method{expand}{CoverageExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{CoverageExperiment} object}

\item{...}{ignored}
}
\value{
a \code{tibble} object
}
\description{
expand
}
\examples{
data(ce)
ce

expand(ce)
}
