% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{aggregate}
\title{Objects exported from other packages}
\value{
Depending on the re-exported function
}
\examples{
1 + 1
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{as_tibble}}}

  \item{S4Vectors}{\code{\link[S4Vectors:aggregate-methods]{aggregate}}}
}}

