% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show-tidyCoverage}
\alias{show-tidyCoverage}
\alias{show,CoverageExperiment-method}
\alias{show,AggregatedCoverage-method}
\title{show method for \code{CoverageExperiment} and \code{AggregatedCoverage} objects}
\arguments{
\item{object}{a \code{CoverageExperiment} or \code{AggregatedCoverage} object}

\item{setup}{a setup object returned from \code{\link[pillar:tbl_format_setup]{pillar::tbl_format_setup()}}.}
}
\value{
\verb{Prints a message to the console describing the contents of the }CoverageExperiment\code{or}AggregatedCoverage` objects.
}
\description{
show method for \code{CoverageExperiment} and \code{AggregatedCoverage} objects
}
\examples{
data(ce)
print(ce)
data(ac)
print(ac)
}
