% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tidyCoverage-package}
\alias{tidyCoverage-package}
\alias{tidyCoverage}
\title{tidyCoverage: Tidyomics-based analysis of coverage tracks over genomic features}
\description{
The tidyCoverage package provides a bridge between Bioconductor
SummarizedExperiment objects and the tidyomics project.
}
\references{
Serizay J, Koszul R., Epigenomics coverage data extraction and aggregation
in R with tidyCoverage. Bioinformatics 40, btae487 (2024).
\doi{10.1093/bioinformatics/btae487}
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/js2264/tidyCoverage}
\item Report bugs at \url{https://github.com/js2264/tidyCoverage/issues}
}
}
\author{
Jacques Serizay
}
\keyword{internal}
