% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageExperiment.R
\name{CoverageExperiment}
\alias{CoverageExperiment}
\alias{CoverageExperiment,list,GRangesList-method}
\alias{CoverageExperiment,list,list-method}
\alias{CoverageExperiment,RleList,GRangesList-method}
\alias{CoverageExperiment,list,GRanges-method}
\alias{CoverageExperiment,RleList,GRanges-method}
\title{CoverageExperiment}
\usage{
\S4method{CoverageExperiment}{list,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{list,list}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{RleList,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{list,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{RleList,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{tracks}{A genomic track imported as a \code{RleList} or a \emph{named} list of
genomic tracks.}

\item{features}{A set of features imported as \code{GRanges} or a \emph{named}
\code{GRangesList}.}

\item{width}{Width to resize each set of genomic features}

\item{scale, center}{Logical, whether to scale and/or center tracks prior to
summarization}

\item{ignore.strand}{Logical, whether to not take the features strand
information}

\item{BPPARAM}{Passed to BiocParallel}
}
\value{
An \code{CoverageExperiment} object
}
\description{
This function initiates a \code{CoverageExperiment} object.
}
\examples{
library(rtracklayer)
library(purrr)
library(plyranges)
features <- system.file("extdata", "TSSs.bed", package = "CoverageExperiment") |> 
    import() |> filter(strand == '+') |> GRangesList(x = _)
tracks <- list(
    RNA_fwd = system.file("extdata", "SRR2045244.fwd.bw", package = "CoverageExperiment"),
    RNA_rev = system.file("extdata", "SRR2045244.rev.bw", package = "CoverageExperiment")
) |> map(import, as = 'Rle')
CE <- CoverageExperiment(tracks, features, width = 5000)
AC
}
