% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageExperiment.R
\name{CoverageExperiment}
\alias{CoverageExperiment}
\alias{CoverageExperiment,BigWigFileList,GRangesList-method}
\alias{CoverageExperiment,BigWigFileList,GRanges-method}
\alias{CoverageExperiment,BigWigFileList,list-method}
\alias{CoverageExperiment,BigWigFile,GRangesList-method}
\alias{CoverageExperiment,BigWigFile,GRanges-method}
\alias{CoverageExperiment,BigWigFile,list-method}
\alias{CoverageExperiment,list,GRangesList-method}
\alias{CoverageExperiment,list,GRanges-method}
\alias{CoverageExperiment,list,list-method}
\alias{CoverageExperiment,RleList,GRangesList-method}
\alias{CoverageExperiment,RleList,GRanges-method}
\alias{CoverageExperiment,RleList,list-method}
\title{CoverageExperiment}
\usage{
\S4method{CoverageExperiment}{BigWigFileList,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{BigWigFileList,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{BigWigFileList,list}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{BigWigFile,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{BigWigFile,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{BigWigFile,list}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{list,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{list,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{list,list}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{RleList,GRangesList}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{RleList,GRanges}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{CoverageExperiment}{RleList,list}(
  tracks,
  features,
  width,
  center = FALSE,
  scale = FALSE,
  ignore.strand = TRUE,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{tracks}{A genomic track imported as a \code{RleList} or a \emph{named} list of
genomic tracks.}

\item{features}{A set of features imported as \code{GRanges} or a \emph{named}
\code{GRangesList}.}

\item{width}{Width to resize each set of genomic features}

\item{scale, center}{Logical, whether to scale and/or center tracks prior to
summarization}

\item{ignore.strand}{Logical, whether to not take the features strand
information}

\item{BPPARAM}{Passed to BiocParallel}
}
\value{
An \code{CoverageExperiment} object
}
\description{
This function initiates a \code{CoverageExperiment} object.
}
\examples{
library(rtracklayer)
library(purrr)
library(plyranges)
TSSs_bed <- system.file("extdata", "TSSs.bed", package = "CoverageExperiment")
features <- import(TSSs_bed) |> filter(strand == '+')

#############################################################################
## 1. Creating a `CoverageExperiment` object from a single BigWigFile
#############################################################################

RNA_fwd <- system.file("extdata", "RNA.fwd.bw", package = "CoverageExperiment")
tracks <- BigWigFile(RNA_fwd)
CoverageExperiment(tracks, features, width = 5000)

#############################################################################
## 2. Creating a `CoverageExperiment` object from a BigWigFileList
#############################################################################

RNA_rev <- system.file("extdata", "RNA.rev.bw", package = "CoverageExperiment")
tracks <- BigWigFileList(list(RNA_fwd = RNA_fwd, RNA_rev = RNA_rev))
CoverageExperiment(tracks, features, width = 5000)

#############################################################################
## 3. Creating a `CoverageExperiment` object from imported bigwig files
#############################################################################

tracks <- list(
    RNA_fwd = system.file("extdata", "RNA.fwd.bw", package = "CoverageExperiment"),
    RNA_rev = system.file("extdata", "RNA.rev.bw", package = "CoverageExperiment")
) |> map(import, as = 'Rle')
CoverageExperiment(tracks, features, width = 5000)

#############################################################################
## 4. Correct for strandness when recovering coverage
#############################################################################

TSSs_bed <- system.file("extdata", "TSSs.bed", package = "CoverageExperiment")
features <- list(
    TSS_fwd = import(TSSs_bed) |> filter(strand == '+'), 
    TSS_rev = import(TSSs_bed) |> filter(strand == '-')
)
tracks <- list(
    RNA_fwd = system.file("extdata", "RNA.fwd.bw", package = "CoverageExperiment"),
    RNA_rev = system.file("extdata", "RNA.rev.bw", package = "CoverageExperiment")
) |> map(import, as = 'Rle')
CoverageExperiment(tracks, features, width = 5000, ignore.strand = FALSE)
}
