% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce}
\alias{ce}
\title{Stranded RNA-seq coverage in yeast}
\format{
\code{CoverageExperiment} object containing 2 features sets and 2 tracks.
}
\usage{
data(ac)
}
\description{
An \code{CoverageExperiment} dataset summarizing stranded RNA-seq coverage
(forward and reverse) over Scc1 peaks and convergent transcription loci
(± 1kb).
Data was generated in yeast (S288c) and aligned to reference R64-1-1.
}
\keyword{datasets}
