% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{ce}
\alias{ac}
\title{Example \code{CoverageExperiment} and \code{AggregatedCoverage} objects}
\format{
\code{CoverageExperiment} object containing 1 features set and 2 tracks.

\code{AggregatedCoverage} object containing 1 features set and 2 tracks.
}
\usage{
data(ce)

data(ac)
}
\description{
Two example objects are provided in the \code{tidyCoverage} package:
\itemize{
\item \code{ce}: a \code{CoverageExperiment} dataset containing stranded RNA-seq coverage
(forward and reverse) over Scc1 peaks (± 1kb).
\item \code{ac}: an \code{AggregatedCoverage} object obtained with \code{aggregate(ce)}.
}
}
\details{
Data was generated in yeast (S288c) and aligned to reference R64-1-1.
}
\keyword{datasets}
