% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/print.R
\name{show,CoverageExperiment-method}
\alias{show,CoverageExperiment-method}
\alias{print.CoverageExperiment}
\title{show}
\usage{
\S4method{show}{CoverageExperiment}(object)

\method{print}{CoverageExperiment}(x, ..., n = NULL)
}
\arguments{
\item{object}{a CoverageExperiment object}

\item{x}{Object to format or print.}

\item{...}{Passed on to \code{\link[=tbl_format_setup]{tbl_format_setup()}}.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than the \code{print_max} \link[pillar:pillar_options]{option}.
Otherwise, will print as many rows as specified by the
\code{print_min} \link[pillar:pillar_options]{option}.}
}
\value{
\code{tibble}

Prints a message to the console describing
the contents of the \code{CoverageExperiment}.
}
\description{
show method for \code{CoverageExperiment} objects
}
\examples{
data(ac)
print(ac)
}
